+++ lib/stories.js
@@ -2,8 +2,6 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import styled from 'styled-components';
 import { RootStyle } from './system';
-import * as themes from './themes';
-import { StoryThemePreview } from './theme-preview';
 
 import * as avatar from './avatar';
 import * as button from './button';
@@ -15,8 +13,6 @@ import * as textInput from './text-input';
 import * as visuallyHidden from './visually-hidden';
 
 const modules = [avatar, button, checkboxButton, icon, loader, mark, textInput, visuallyHidden];
-
-const themeOptions = Object.entries(themes).map(([name, theme]) => ({ id: name, label: name }));
 
 const Select = ({ value, options, onChange }) => (
   <select value={value} onChange={(e) => onChange(e.target.value, e)}>
