+++ lib/system.js
@@ -1,5 +1,6 @@
 import React from './react';
 import styled, { createGlobalStyle } from 'styled-components';
+import * as themes from './themes';
 
 function* themeToCSSVars(theme, prefix = '--') {
   for (const [key, value] of Object.entries(theme)) {
@@ -54,6 +55,26 @@ for (const size of sizeNames) {
   sizes[size] = `font-size: var(--fontSizes-${size});`;
 }
 
+const themeOptions = Object.entries(themes).map(([name, theme]) => ({ id: name, label: name }));
+
 export const Story_style_providers = () => {
   
 }
+
+const  = () => {
+  const [value, onChange] = React.useState(themeOptions[0].id);
+  
+  return (
+    <>
+      <SectionH>Themes</SectionH>
+      <p>
+        This library includes a set of themes 
+      </p>
+      <RootStyle theme={themes[value]} />
+      <Select value={value} onChange={(val) => onChange(val)} options={themeOptions} />
+      <StoryThemePreview />
+    </>
+  )
+}
+
+expo
