+++ lib/theme-preview.js
@@ -4,15 +4,6 @@ import styled from 'styled-components';
 import { RootStyle, LocalStyle } from './system';
 import * as themes from './theme';
 
-const ThemeContainer = styled.div`
-  ${setThemeVars}
-  font-family: var(--fonts-sans);
-  font-size: 100%;
-  color: var(--colors-primary);
-  --local-colors-secondary: var(--colors-secondary);
-  background-color: var(--colors-background);
-`
-
 const Grid = styled.div`
   display: grid;
   grid-gap: var(--space-1);
@@ -37,7 +28,7 @@ const createCells = (theme) => {
 }
 
 export const ThemePreview = ({ theme }) => (
-  <ThemeContainer theme={theme}>
+  <LocalStyle theme={theme}>
     {['background', 'secondaryBackground'].map(bg => (
       <div key={bg} style={{ backgroundColor: theme.colors[bg], padding: 'var(--space-1)' }}>
         <h3>
@@ -57,7 +48,7 @@ export const ThemePreview = ({ theme }) => (
         </Grid>
       </div>
     ))}
-  </ThemeContainer>
+  </LocalStyle>
 )
 
 const themeItems = Object.entries(themes).map(([id, theme]) => ({ id, theme }))
@@ -67,23 +58,6 @@ const ThemeGrid = styled.div`
   grid-gap: var(--space-1);
   grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
 `
-
-const  = ({ themeOptions }) => {
-  const [value, onChange] = React.useState(themeOptions[0].id);
-  
-  return (
-    <>
-      <SectionH>Themes</SectionH>
-      <p>
-        This library includes a set of themes 
-      </p>
-      <RootStyle theme={themes[value]} />
-      <Select value={value} onChange={(val) => onChange(val)} options={themeOptions} />
-      <StoryThemePreview />
-    </>
-  )
-}
-
 export const StoryThemePreview = () => (
   <ThemeGrid>
     {themeItems.map(({ id, theme }) => (
